docs. move utils to the optimizer module.
